Fix file paths and include statements
