+++ config/config.json
@@ -1,0 +1,13 @@
+{
+    "env": {
+        "ODPT_CONSUMER_KEY": "2bbf82cd3d5647265c321eab3cd84ef0178ec8f4b472bf6655d13f4f11f92c67"
+    },
+    "mode": {
+        "dijkstra": false,
+        "preprocess": true
+    },
+    "option": {
+        "api": true,
+        "test": false
+    }
+}
