+++ docs/w7.md
@@ -1,0 +1,1 @@
+A(HH:MM) -> B(HH:MM)
