+++ include/dijkstra.h
@@ -1,0 +1,11 @@
+#pragma once
+#include <vector>
+#include <iostream>
+
+#include "model.h"
+
+namespace dijkstra
+{
+
+    int main();
+}
