+++ src/main.cpp
@@ -1,34 +1,57 @@
-#include "preprocess.h"
-// #include "dijkstra.cpp"
+#include "../include/preprocess.h"
+#include "../include/dijkstra.h"
+#include "../include/config.h"
 
 using namespace std;
 
 int main(int argc, char *argv[])
 {
-    if (argc < 2)
+    // if (argc < 2)
+    // {
+    //     cout << "Usage: xtrains.exe <command>" << endl;
+    //     cout << "Commands:" << endl;
+    //     cout << "  preprocess" << endl;
+    //     cout << "  dijkstra" << endl;
+    //     return 1;
+    // }
+
+    // string command = argv[1];
+    // vector<string> args;
+    // for (int i = 2; i < argc; i++)
+    // {
+    //     args.push_back(argv[i]);
+    // }
+
+    // if (command == "preprocess")
+    // {
+    //     return preprocess::main(argc - 2, argv + 2);
+    // }
+    // else if (command == "dijkstra")
+    // {
+    //     return dijkstra::main();
+    // }
+    // else
+    // {
+    //     cout << "Unknown command." << endl;
+    //     return 1;
+    // }
+
+    bool dijkstra = readMode("dijkstra");
+    bool preprocess = readMode("preprocess");
+
+    if (dijkstra)
     {
-        cout << "Usage: xtrains.exe <command>" << endl;
-        cout << "Commands:" << endl;
-        cout << "  preprocess" << endl;
-        cout << "  dijkstra" << endl;
-        return 1;
+        return dijkstra::main();
     }
-
-    string command = argv[1];
-
-    if (command == "preprocess")
+    else if (preprocess)
     {
         return preprocess::main();
     }
-    else if (command == "dijkstra")
-    {
-        cout << "NOT Implemented." << endl;
-    }
     else
     {
-        cout << "Unknown command." << endl;
+        cout << "No mode specified." << endl;
         return 1;
     }
 
     return 0;
-}
+} // namespace preprocess
