+++ src/preprocess.cpp
@@ -1,8 +1,7 @@
-#include "preprocess.h"
+#include "../include/preprocess.h"
 
 namespace preprocess
 {
-    using namespace std;
     using json = nlohmann::json;
 
     int timeToSeconds(const std::string &time)
@@ -27,7 +26,7 @@ namespace preprocess
     {
         api::URLBuilder builder(api::Endpoint::Railway);
         builder.addQuery(api::Query::Operator, api::operator_map[api::Operator::TokyoMetro]);
-        string res = api::getRequest(builder);
+        std::string res = api::getRequest(builder);
         json j = json::parse(res);
         railway::DB railwayDB;
         int railwayID = 0;
@@ -35,8 +34,8 @@ namespace preprocess
         {
             if (railway.contains("odpt:lineCode") && railway.contains("owl:sameAs"))
             {
-                string railwayCode = railway["odpt:lineCode"].get<string>();
-                string railwayLongID = railway["owl:sameAs"].get<string>();
+                std::string railwayCode = railway["odpt:lineCode"].get<std::string>();
+                std::string railwayLongID = railway["owl:sameAs"].get<std::string>();
                 railwayDB.add({railwayID, 0}, {railwayCode, railwayLongID});
             }
             railwayID++;
@@ -54,15 +53,15 @@ namespace preprocess
         {
             api::URLBuilder builder(api::Endpoint::Station);
             builder.addQuery(api::Query::Railway, v.long_id);
-            string res = api::getRequest(builder);
+            std::string res = api::getRequest(builder);
             json j = json::parse(res);
             for (const auto &station : j)
             {
                 if (station.contains("odpt:stationCode") && station.contains("owl:sameAs"))
                 {
-                    string stationCode = station["odpt:stationCode"].get<string>();
-                    string stationID = station["owl:sameAs"].get<string>();
-                    string stationCodeNumStr = stationCode.substr(stationCode.size() - 2);
+                    std::string stationCode = station["odpt:stationCode"].get<std::string>();
+                    std::string stationID = station["owl:sameAs"].get<std::string>();
+                    std::string stationCodeNumStr = stationCode.substr(stationCode.size() - 2);
                     int stationCodeNum = stoi(stationCodeNumStr);
                     stationDB.add({k.first, stationCodeNum}, {stationCode, stationID});
                 }
@@ -78,7 +77,7 @@ namespace preprocess
         station::DB stationDB;
         railway::DB railwayDB;
         train::DB trainDB;
-        vector<api::Calendar> calendars = {
+        std::vector<api::Calendar> calendars = {
             api::Calendar::Weekday,
         };
 
@@ -89,45 +88,37 @@ namespace preprocess
                 api::URLBuilder builder(api::Endpoint::TrainTimetable);
                 builder.addQuery(api::Query::Railway, v.long_id);
                 builder.addQuery(api::Query::Calendar, api::calendar_map[calendar]);
-                string res = api::getRequest(builder);
+                std::string res = api::getRequest(builder);
                 json j = json::parse(res);
                 int trainID = 0;
                 for (const auto &train : j)
                 {
                     if (train.contains("odpt:trainNumber") && train.contains("odpt:trainTimetableObject"))
                     {
-                        string trainNumber = train["odpt:trainNumber"].get<string>();
-                        vector<train::Stop> stops;
-                        // printf("%s ", trainNumber.c_str());
+                        std::string trainNumber = train["odpt:trainNumber"].get<std::string>();
+                        std::vector<train::Stop> stops;
                         json trainTimetable = train["odpt:trainTimetableObject"];
                         for (const auto &station : trainTimetable)
                         {
                             if (station.contains("odpt:departureStation") && station.contains("odpt:departureTime"))
                             {
-                                string time = station["odpt:departureTime"].get<string>();
+                                std::string time = station["odpt:departureTime"].get<std::string>();
                                 int seconds = timeToSeconds(time);
-                                string stationLongID = station["odpt:departureStation"].get<string>();
+                                std::string stationLongID = station["odpt:departureStation"].get<std::string>();
                                 const auto &[sk, sv] = stationDB.read("long", stationLongID);
-                                // printf("%s %s", v.short_id.c_str(), v.long_id.c_str());
-                                // printf("%s %d ", v.short_id.c_str(), seconds);
-                                // printf("%d %d %d  ", k.first, k.second, seconds);
                                 stops.push_back({sk.first, sk.second, seconds});
                             }
                             else if (station.contains("odpt:arrivalStation") && station.contains("odpt:arrivalTime"))
                             {
-                                string time = station["odpt:arrivalTime"].get<string>();
+                                std::string time = station["odpt:arrivalTime"].get<std::string>();
                                 int seconds = timeToSeconds(time);
-                                string stationLongID = station["odpt:arrivalStation"].get<string>();
+                                std::string stationLongID = station["odpt:arrivalStation"].get<std::string>();
                                 const auto &[sk, sv] = stationDB.read("long", stationLongID);
-                                // printf("%s %s", v.short_id.c_str(), v.long_id.c_str());
-                                // printf("%s %d ", v.short_id.c_str(), seconds);
-                                // printf("%d %d %d  ", k.first, k.second, seconds);
                                 stops.push_back({sk.first, sk.second, seconds});
                             }
                         }
                         trainDB.add({k.first, trainID}, {trainNumber, stops});
                     }
-                    // printf("\n");
                     trainID++;
                 }
             }
@@ -138,15 +129,37 @@ namespace preprocess
 
     int main()
     {
-        // railways();
-        railway::test();
+        // check --api option
+        // bool requireAPI = false;
+        // bool test = false;
+        // for (int i = 0; i < argc; i++)
+        // {
+        //     std::string arg = argv[i];
+        //     if (arg == "--api" || arg == "-A")
+        //     {
+        //         requireAPI = true;
+        //     }
+        //     else if (arg == "--test" || arg == "-T")
+        //     {
+        //         test = true;
+        //     }
+        // }
 
-        // stations();
-        station::test();
+        bool requireAPI = readOption("api");
+        bool test = readOption("test");
 
-        // trains();
-        train::test();
-
+        if (requireAPI)
+        {
+            railways();
+            stations();
+            trains();
+        }
+        if (test)
+        {
+            railway::test();
+            station::test();
+            train::test();
+        }
         return 0;
     }
