--- src/station.cpp
@@ -1,8 +1,0 @@
-#include "station.h"
-
-namespace station
-{
-    using namespace std;
-
-    
-} // namespace station
